Include an HTML sanitiser in Govspeak.
All applications which use Govspeak need to be able to defend
against malicious XSS and content-based attacks. As a first step,
I've moved the `HtmlValidator` class from `govuk_content_models` to
the `govspeak` gem. However, at this point it's still the
responsibility of the individual applications to use this validator.