Add a $DownloadLink{} extension for the new download link callouts
These callouts require the existance of file size and file format to
move away from URLs including this metadata (as they currently do).

The suggested markdown extension has been discussed with members of
the content team and they agree that it's understandable.

NOTE: This work came about as part of a spike and may not proceed
past code review. It's just an example to show what could be done.