Parse legislative and priority lists correctly
When omitting a new line before certain list extensions, the markdown
parser has difficulty deciding where to close the tags and was rendering
out the ol opening and closing tag as text inside a <p> if there wasn't
a newline after the text. This commit only parses it if it's at the
start of a document or with two new lines preceding it.