Upgrade sanitize dependency to 2.1.0
The older version used by govspeak (2.0.3) contains a bug where colons in
ids would not be allowed, which prevents footnotes being saved as part
of a document body in specialist-publisher.

This was fixed in 2.1.0 (see
https://github.com/rgrove/sanitize/pull/87).

Trying to force an upgraded version of sanitize to be activated using
the specialist-publisher Gemfile is a non-starter as bundle refuses to
upgrade to that version as it's too far ahead of Govspeak's
requirements.