Update to latest version of kramdown gem
Note that this required some slight changes to the tests due to subtle
changes in kramdown's behaviour:

* There is a bug in kramdown that means &amp; is not correctly converted
to it's numeric equivilant.
* three full stops now get converted to a &hellip; in newer versions of
kramdown so the test strings needed updating to better reflect real
usage.

A bug report has been filed with kramdown on the conversion of certain
entities to their numeric equivilance. This is not directly a problem
for our usage as we do not use :numeric output, only symbolic.