Allow table cells to have a text alignment
Kramdown has a markdown pattern for allowing left, centred and right
aligned cell contents:
http://kramdown.gettalong.org/quickref.html#tables

To achieve this it uses a style property, eg `style=‘text-align: right’`

We don’t want to allow any styles to be applied, just the subset that
Kramdown uses. Instead, whitelist the attribute but use a Sanitise
transformer to remove the style unless it matches the 3 possible values
it can have.