Add failing test for external links with IALs
Inline attribute lists (IALs) were added to links in Kramdown 1.6.0:
https://github.com/gettalong/kramdown/commit/627978525cf5ee5b290d8a1b867
5aae9cc9e2934
http://kramdown.gettalong.org/syntax.html#block-ials
https://github.com/gettalong/kramdown/blob/60795eb95f1588d41e105407add2d
3ba6a7fc101/doc/news/release_1_6_0.page

This changed the `add_link` signature. We monkey patch this in
`kramdown_with_automatic_external_links` to add “rel=external” but we
didn’t update that signature when we bumped Kramdown. There were no
tests using the IAL feature so the error was not spotted. Live markdown
examples do however trigger this feature, which leads to the error:
`ArgumentError: wrong number of arguments (5 for 3..4)`

* Include simple IAL external link test
* Include kramdown’s own IAL link tests