Blockquote extra quote removal
Importing the blockquote extra removal class from Whitehall. The tests
have been maintained as per those in Whitehall.

It does look like there are likely a number of problems within this
function and tests:

- It likely catches items on a single line which would not be rendered
  in a blockquote if there is not a line break before them
- It seems to handle quotes within the blockquote outside of the
  condition here poorly. In a number of cases it looks like tests have
  been planned for them to work properly but wasn't implemented.

Due to this being planned for consistency with Whitehall I'm reluctant
to change behaviour here in case that causes backwards compatibility
issues.