Remove entity output option
Prior to post-processing being introduced to govspeak we had the ability
to specify the entity_output of the HTML. By using nokogiri to format
to format the HTML we lose the ability to do this as nokogiri will
automatically perform conversions to the entities.

In cases where nokogiri encounters a encoded HTML entity that it can
safely output as a utf-8 character it will automatically replace the
entity with the character. For example: `&yen;` is converted to `¥`

In cases where nokogiri encounters a numeric HTML entity it will convert
that into a symbolic one. For example: `&#62;` is converted to `&gt;`

I haven't found any instances where our applications are using the
`:numeric` option and believe it was only introduced as an option as a
side effect of a different change, rather than being something that was
required:
https://github.com/alphagov/govspeak/commit/74ea8c77293ffa2ef7c2876e1b832072ae297017