embed:link markdown extension
This adds support for `embed:link:<content_id>` within govspeak. For
example if a content item wanted to link to a content item that has a
content id of `901c17a5-67ad-4747-aef6-b17e92f73948` the following syntax
would be used: `[embed:link:901c17a5-67ad-4747-aef6-b17e92f73948]`.

When the document is rendered this syntax would convert to one of 3
things:
1) `"<a href="{url}">{title}</a>"` - where link title and url is known
2) `"{title}"` - where link title is known but url is not (the content item
might be unpublished)
3) `""` - where no link details are available

Links are passed into the Govspeak rendering in the form of an array of
struct objects, where they can have a content_id, a title and a url
field. When a content item cannot be linked to (for instance only in
draft form or unpublished with gone type).