Loosen ActionView dep for Rails 5 compatibility.
This allows the gem to be included in apps that use Rails 5.

We test against different versions of Ruby, but ActionView 5 does not
build in versions before 2.3. Because the `>=` gem specifier always
finds the latest version, we need to use a different specifier for
earlier Ruby versions, so this configures Travis to use a different
Gemfile for each one.