Render bad usage of markdown as empty strings
Prior to this change `[embed:attachments:filename.jpg]` would be
rendered on the frontend amongst the content. This change means that it
would be rendered as an empty string.

Although it is debatable which route is better, having this render empty
strings does allow us to resolve some issues with missing attachments on
Specialist Publisher.