Convert regex matches to equals
This changes how we check whether a content_id matches for
link/contacts/attachments. As the content_id is user input it should not be
treated as a regex, so the equality operator is used instead. This
shouldn't break any existing usages as nothing should have been reliant
on regexs.

This is to fix an error that was occurring of: "premature end of
char-class: /1-1992 G-BJRT[2/"