Handle links that have a blank or missing href
If the markdown contains an empty markdown link `[foo]()` or it contains
`a` tags with missing or empty `href` attributes we should strip them out
of the list of hrefs we return.  They're not valid links so there's no
point in checking them.  We use CSS selectors on the document to strip
out any links that have no href attribute (which means we ignore valid
`a` tags like `<a name="table-of-contents">ToC</a>`), but we use ruby to
filter out blank `href` attributes because it's more flexible.