Remove WithDeepMerge and use Sanitize::Config.merge
Sanitize::Config objects are now frozen by default so we can't manipulate
them in place.  It does now come with a `merge` method that lets you
combine them though, and this means we no longer need our `deep_merge`
method.

While we're manipulating the config we can also remove some of the
settings that we've added because they have been rolled into the default
RELAXED settings.  Specifically: `div`, `aside`, `span` are now allowed
elements, `id`, `class` are allowed attributes, and `rel` is allowed on
an `a` tag.