Remove unreliable test
We can't rely on this test as we get different results based on what
version of ActiveSupport we use. With < 5.2.1 we get an alt tag for an
image and for later versions we do have one.

I feel like most of the things tested here are covered in other tests
and this full HTML output is a bit overkill since it relies on libraries
not covered by this code.