Render contacts from a schema like format
Prior to this PR the data for contacts inside was based on the needs of
the HTML that would be rendered. This makes it rather incompatible with
the contact format that is used [1].

This changes the expected data for contacts to be based on the
content schema - so that a Content Item returned from Publishing API get
contact (or from Content Store) can be accepted.

It also drops a worldwide office link which is a tricky to implement
feature that Whitehall does not seem to actually use as part of govspeak
rendering:

Whitehall uses this code to render Contacts: https://github.com/alphagov/whitehall/blob/58374b3ba0de07be9dfef269028e1663f4fda0f8/app/helpers/govspeak_helper.rb#L144-L148
The link is only rendered if the Contact is a WorldwideOffice: https://github.com/alphagov/whitehall/blob/235a26c29c18b081a23a075d76f0f2ef34c4e360/app/views/contacts/_contact.html.erb#L47-L49
Worldwide office can pretend to be a Contact:
https://github.com/alphagov/whitehall/blob/235a26c29c18b081a23a075d76f0f2ef34c4e360/app/models/worldwide_office.rb#L18-L22
but will never be loaded from a Contact.find(x) method as per the first
point.

[1]: https://github.com/alphagov/govuk-content-schemas/blob/4e976c696ddd6ebfd1c4e1da52d6d5af54c79ba4/formats/contact.jsonnet