Change way we load in locale files
The previous method did not work when this gem was used in a ruby app as
the load method was relative to the current directory. This changes this
to load the files relative to the file.

This also loads the locale files in when the gem starts rather than at
app runtime. This is so the gem does not surprise apps by adding locale
information later that can not be overriden by apps.