Load required files using Govspeak.root for path
This creates a method on the Govspeak module that will return the root
path of the Gem. This is to make it simpler to include files this gem
requires.

The relative loading of files prior to this meant files were only loaded
when running the gems tests and not when the gem is used in an
application.