Don't render empty address or UK country name
This re-applies a quirk of Whitehall where United Kingdom as a country
name is not shown in any contact addresses that are shown to users as
it's considered implicit. It feels a bit weird to do this here and
potentially frustrating in some contexts but it stops us introducing a
change.

This also adds code to deal with a common problem that contacts have an
array of empty data for postal_addresses in the Publishing API

As an example (from Publishing API):

Document.find_by(content_id: "fb6777d8-1e19-427f-8dae-0ef50f0ff359").live

```
details:
  {"description"=>"",
   "contact_type"=>"General contact",
   "title"=>"Visa Enquiries",
   "contact_form_links"=>[{"title"=>"Visa Enquiries", "link"=>"", "description"=>""}],
   "post_addresses"=>
    [{"title"=>"",
      "street_address"=>"",
      "locality"=>"",
      "region"=>"",
      "postal_code"=>"",
      "world_location"=>""}],
   "email_addresses"=>
    [{"title"=>"",
      "email"=>"All enquiries via the UK Visas website at www.gov.uk/visas-immigration"}],
   "phone_numbers"=>[{"title"=>"Visa Enquiries", "number"=>"00 44 203 481 1736"}]},
```

(No comment on that email address)