More protection against empty fields
This is a bit of sad commit as I didn't really want to add this stuff
in. This adds more protection against empty data from the Publishing API
where it returns fields with empty strings.

I'm hoping to fix most of the data Whitehall has currently but I don't
think I'll fix it all so this protection will probably be needed
ongoing.