Don't render [Image: <file>] that is within a paragraph
Doing this causes invalid HTML (it embeds elements such as figure, div
and p) and undesirable visual presentation - raw HTML is output.

Since this tag is not used in the wild just yet, it can be fixed before
users make use of it.

These same problems apply to the other Image embeds however I dare not
change those in case there are places where users are relying on the
invalid HTML is produced and it breaks documents.