Allow sanitize option on initialize
This allows sanitizing the source that is given to Govspeak::Document.
This is to strip any HTML that is forbidden from the input before it is
processed. With the addition of this the intention is to remove two
previous methods: `to_sanitized_html` and `to_sanitized_html_without_images`.
Both of these methods were rather dubious as they sanitized HTML after
the processing which meant that none of the govspeak tests ran against
this code.