Remove to_sanitized_html and to_sanitized_html_without_images methods
The `to_sanitized_html_without_images` method was only used by the
content planner application which is long retired.

The `to_sanitized_html` method has been removed in favour of a `sanitize`
option given on initialize. This approach is considered advantageous as
it does the sanitising before the govspeak is processed rather than
after. This resolves a problem that any custom extensions could be
escaped by the sanitising process after processing the govspeak.
This is also consistent with the approach of the `valid?` method.