Render embed:attachment in post processing
This is a step towards turning HTML sanitisation on by default in
govspeak. The embed attachment is the only govspeak transformation that
produces HTML that is altered by the HTML sanitiser (which feels rather
lucky).

The approach I've taken to allow this is to convert the item to a custom
element which can be whitelisted by the sanitizer and then performing
the transformation in post processing.

I hit some issues when treating the HTML as a self closing tag
`<govspeak-embed-attachment content-id="blah">` would swallow subsequent
elements. And `<govspeak-embed-attachment content-id="blah" />` seemed
to create an extra new line. I think these all stem from the nokogiri
processing switching this element to be
`<govspeak-embed-attachment></govspeak-embed-attachment>` so it seems
to avoid surprises by just doing this in the first place.