Sanitize input by default
This turns on HTML sanitization by default for all input into Govspeak.
This is to take a security by default approach as it's not clear that by
default you could include script in govspeak - particularly as this is
the primary syntax for inputting content on GOV.UK

I expect the impact of this change to be low as most publishing apps use
the `valid?` method on govspeak to check input, and we don't expect HTML
to be entered by users. Most likely this may affect parts of apps that
are using govspeak for simple templating - for instance Content
Publisher will have a problem where it uses target blank attribute on
some links.

The trickiest part of implementing this was finding a safe place to
perform the sanitization. If the sanitization is performed on the raw
user input then blockquote markdown syntax is broken and the valid?
method won't work. I decided the best point was to do it after all
kramdown processing was done and the post-processing tasks can be used
for any transformations that involve creating HTML outside of the
sanitize config (as was the case with embed:attachment).