Iterate blockquote remover
This fixes an issue that was found where a space after the start of a
quote is not matched by the expression.

I switched this to be two regexes that match strings that either start
with a quote or end in a quote. This replaces the optional approach
where either quote is optional and this effectively matched all block
quote strings - this was hard to maintain because it ended up changing
lines with just "> ".