Create AttachmentLink syntax
Like Attachment, this is marked as experimental as it is intended to be
used in developing inline attachments for Content Publisher and may be
subject to change.

The likelihood of this one changing is much lower than Attachment but it
seemed worth treating it in the same way for the time being.