Remove embed:attachment extension
This has never worked properly and is not used by any apps. The
intention is for this to be replaced with the `[Attachment:id]`
extension which will use govuk_publishing_components for rendering.

Removing this allows a lot of the logic from the attachment presenter to
be removed as well as the dependency on the Money gem.