Govspeak attachment links are rendered inside a paragraph
This adds a test that attachment links are rendered in a paragraph.

To ensure we still return an empty string if a attachment link doesn't
exist this checks whether the attachment exists before returning the
custom HTML element, otherwise this would end up returning an empty
paragraph element.

This applies the same behaviour to a block attachment so these are not
entered as custom elements without an attachment existing.