Remove experimental status of Attachment extensions
These are now in production use by Content Publisher so their
experimental status has been removed.

As part of this the `[embed:attachments:inline:content-id]` extension
has been deprecated as this provides the same functionality as
`[AttachmentLink:attachment-id]`. At the time of writing this extension
is not directly used by any government publishers and is only used
behind the scenes in Specialist Publisher - A user enters a numeric
syntax and at the point of sending to Publishing API the govspeak is
converted to the inline syntax.