Simplify govspeak table tests
There was previously a variety of different tests used to check
properties of tables that may have called govspeak problems in
Whitehall. Of these I think the only one of these we should keep long
term is the check that presence, or lack of, `<tbody>` elements.

I did consider keeping the scenario of quotes - but as that just looks
like valid HTML I don't think it's necessary.