Remove forbidden unicode characters from output
This is to resolve a blocking issue that prevents Whitehall from
upgrading to Govspeak > 5.5. A problem we have is that a number of pieces of
Whitehall content have unusual unicode characters that the Santize gem
strips out, this then makes the content fail HTML Validation and was
introduced as part of a Santize update in
https://github.com/alphagov/govspeak/pull/127.

The approach I've taken to resolve this is to remove forbidden unicode
characters from the source markdown that is converted to HTML. Since
these characters are invalid in a HTML output it seems reasonable to
consider them as invalid as part of govspeak input.

Looking through Whitehalls database I found there to be 765
Edition::Translation models and 248 GovspeakContent models that have at
last one of these characters in their body fields.

Initially I intended to resolve this with a data migration on Whitehall,
however that would have caused a problem if anyone used these characters
again as it would be very difficult to determine why content was failing
govspeak validation. I figure it's actually more user friendly to strip
the character out and then if they notice a space or similar is missing
they can correct that with their keyboard.

I noticed that in 2019 there had been 74 of these items added so this
doesn't appear to be a legacy issue. My current theory is that perhaps
these are characters that are present in content that is being pasted
from such as a PDF document.

Anyway, with this merged in Whitehall can proceed to be upgraded to
Govspeak 6 as with this and
https://github.com/alphagov/govspeak/pull/159 all GovspeakContent models
are valid govspeak and only 14 Edition::Translation bodies are invalid,
all of which are also invalid under Govspeak 5.5.