Change the characters that we test for sanitization
The gem that we use to sanitize html has recently changed the list of characters that it checks for.
https://github.com/rgrove/sanitize/commit/0d4158fe3108709cfe93c40371cf01e07b8debf1

This changes the tests that we run to ensure that characters are being stripped to reflect these changes.

This change is being made prior to introduction of rubocop govuk. I've had to remove govuk-lint in order to get this to pass but will be raising a PR shortly that introduces rubocop-govuk.