Add Brexit CTA extension
When $BrexitCTA is used in a document then we replace it with a CTA block of
HTML with static content. This has been introduced with the intention of
reducing the need for content designers to manually change this content each
time Brexit messaging changes. When the messaging does change, a developer will
need to update the extension content and release a new version of the gem to
make it available to publishing apps.