Modify version number for actionview
This is part of some work to upgrade content publisher to rails 6.

Trello - https://trello.com/c/g9LIshqc/1158-upgrade-content-publisher-to-rails-6

In order for content publisher to be upgraded it needs to use a higher version of actionview than govspeak would allow. This allows versions 5 and 6 of actionview which should mean that we can continue upgrading content publisher and other apps that will require this.