Prevent links in table headers being stripped
The substitution that strips # from table headers is effectively
taking a copy of the text but not the child html which means that
links and other declared-safe html is lost. We know the # that
indicates a header will always appear at the start and therefore be
in the first child so this targets the substitution to the content of
the first child instead of the whole content. Tests added to confirm
expected behaviour persists.