Customise footnote markup
The default footnote "backlink" only has an arrow as content. Screen
readers only say "right arrow curving left" when they get to this element.
This is a problem in terms as accessibility as it is unclear to the
users what those links do. A better solution is to use text to indicate
what the link does, which is what this is aiming to achieve.

This commit changes the footnote markers from being just numbers to
[footnote 1,2,3...i]. Similarly, it replaces the "back link" which is
not very helpful to the user, with "[go to where this is referenced]",
which describes precisely what the link should do.