Take an optimistic approach to versioning govuk_publishing_components
This doesn't specify a maximum version for govuk_publishing_components
which is done so that we don't have to re-release this gem each time a
new version of govuk_publishing_components is released.

As this gem only uses 2 components it seems unlikely that new versions
of govuk_publishing_components would break this gem unintentionally and,
given the frontend nature of the components, we'd likely need more than
our minitest tests to confirm compatibility for a breaking change.

If it turns out that govuk_publishing_components becomes more volatile
in it's changes we may need to switch back to pessimistic versioning but
I'm expecting this will just save people a chore moving forward.