Fix issue rendering $CTA followed by $C block
There was a issue were the $C in the first $CTA tag was being matched
with the final $C tag. This cause the block of content to be rendered as
a contact instead of a call to action. This changes the priority of the
regexes evaluated so $CTA tags identified and parsed before $C,
preventing the mis-match.