Fix tests for newer versions of govuk-publishing-components
A recent change to govuk_publishing_components is to set display none
for certain styles in print contexts [1]. These unfortunately broke a
number of tests in this gem as they added extra class elements that are
not in the markup of our assertions.

I've resolved this for attachments by truncating what we assert on so
only the first part of the class name is checked. This seems to work but
is naturally a bit fragile.

For the Start button elements I've temporarily added in the missing
class so that the tests pass. However I think it's a little too fragile
to have govuk_publishing_components HTML included verbatim so I intend
to clean this up on the next commit.

[1]: https://github.com/alphagov/govuk_publishing_components/commit/6fe73e6a519e4a097182ab1192d9eccc91232823