Reduce button test verbosity
These tests were quite fragile to any changes in
govuk_publishing_components as they expected the HTML from that gem to
be identical to assertions.

To try reduce fragility I have added a new assert selector method that can
match a CSS selector. This then allows us to reduce the amount of things
we assert on to reduce the coupling with the gem.

In doing this I noticed that a few of the tests are basically repeats of
earlier tests so removed them. I also simplified the register to vote
test to not include the start button SVG to make that a little less
fragile, I did this because I couldn't think of a nice quick way to
assert on everything it is doing.