Fix footnote functionality in Legislative lists
We currently don't have the ability to render footnotes in Legislative
lists. This workaround substitutes the footnote anchor links (e.g. [^1]
for it's equivalent html.

In order to know whether the document contains legislative lists
with footnotes we have to check for this during the preprocess stage.
The markdown manually converted to html is then appended or
substituted to the html during the document's post process stage.