Upgrade Sanitize to 6 to drop Nokogumbo
This version of Sanitize no longer has a dependency on Nokogumbo and
will stop our apps emitting a Nokogumbo warning.

Although Sanitize 5.2.3 -> 6 is a major version upgrade, the changes are
very benign. With the bump given a major status due to the dropping of
old Ruby versions (that this gem had already dropped) [1]

[1]: https://github.com/rgrove/sanitize/blob/main/HISTORY.md#600-2021-08-03