Reverts acronym change
This code is buggy as it replaces abbreviations in HTML tag attributes,
rather than just text tags. This has knock on consequences of removing
HTML attachments from documents in Whitehall when creating a new
edition.

This commit reverts the change for now with a view to updating the
logic in the near future.

Zendesk:
https://govuk.zendesk.com/agent/tickets/4819383