Remove `simplecov-rcov` code coverage gem
With the move to Ruby 2.7, deprecation notices started appearing from the `simplecov-rcov` gem. This gem seems to have been abandoned – the [current version][1] is 0.2.3, released in 2011.

It seems that this gem was being used to enable code coverage reporting in the Hudson (now known as Jenkins) CI user interface.

Having discussed it with other developers on the team, it seems like the best approach is to simply remove this gem. Its inclusion in this codebase is [unconventional][2] as it doesn't seem to be actively used in other GOV.UK applications, nor does the Jenkins coverage report seem to be used by developers working on this codebase.

[1]: https://rubygems.org/gems/simplecov-rcov
[2]: https://docs.publishing.service.gov.uk/manual/conventions-for-rails-applications.html