Fix uninitialized constant ActiveSupport::XmlMini::IsolatedExecutionState
After the move to Rails 7, the following error was causing tests to fail:
```
uninitialized constant ActiveSupport::XmlMini::IsolatedExecutionState
```

The solution was to explicitly require `active_support` before requiring the Hash and Array core extensions. This is the suggested approach given in the [Active Support documentation][1], but presumably the previous combination of Ruby and Rails was more forgiving of its omittance.

This change is based on the solution identified here:
https://github.com/rails/rails/issues/43851#issuecomment-1001921207

[1]: https://guides.rubyonrails.org/active_support_core_extensions.html#stand-alone-active-support