Use a composite job to satisfy branch protections
This commit implements a 'composite job' in GitHub Actions to satisfy the branch protection rules which require that the 'test' check has passed.

We use a matrix job to run the test suite in multiple Ruby versions. Each variant in the matrix (i.e. each Ruby version) spawns a new check with its own name – e.g. 'test (2.7)', 'test (3.0)' and so on. This meant that there wasn't one sinlge check which the branch protection rules could mark as a dependency.