Allow aria-hidden to be present in SVGs
A recent change to [govuk_publishing_components](https://github.com/alphagov/govuk_publishing_components/pull/2533/commits/0133558abe3547ef9f0cb2bf5f699d7c63c6d7fc)
changes the SVG used in buttons to have the attribute `aria-hidden="true"` instead of
`role="presentation"` so it is not flagged to screen readers.

This causes the validation of govspeak to fail since it is not expecting
the `aria-hidden` value to be there.

To prevent this we have allowed `aria-hidden` to be be in the svg.