Don't validate changes made by post-processors
The HTML Validator was sanitizing Markdown which had been converted to HTML and had post-processors applied. This meant that any unsafe HTML added by a post-processor was causing the source Markdown to be considered invalid.

Logically, it seems more sensible for HTML to be validated after conversion from Markdown, but **before** the post-processors have altered the HTML.

This way, we only validate user input rather than validating the output of our own post-processors, which we consider to be trusted.