Ensure html attachments without <tbody> elements duplicate successfully
At the moment, there's a bug where some attachments are not being
successfully duped when a new edition is created by the
`Edition#create_draft` method.

After some digging it's become clear that this occurs when a user inputs
questionable HMTL for tables. Specifically when they don't use a
<tbody> for the body of the table.

When a user saves a HTML attachment on Whitehall it's validates it via
the Govspeak::HtmlValidator.

This calls the Document#to_html method. When we pass `sanitize: true`
into the options we then call the `Sanitize#sanitize` method from the
`Sanitize` gem which uses the `fragment` method. Here's the code:

```
def fragment(html)
  return '' unless html

  frag = Nokogiri::HTML5.fragment(preprocess(html), **@config[:parser_options])
  node!(frag)
  to_html(frag)
end
```
https://github.com/rgrove/sanitize/blob/main/lib/sanitize.rb#L66-L68

This `Nokogiri::HTML5.fragment` has some cool, but slightly unexpected
behaviour that is outlined here
https://makandracards.com/makandra/481802-how-to-prevent-nokogiri-from-fixing-invalid-html

Essentially though it fixes up invalid HTML. One side effect is that
it can add extra linebreaks that were not there before, which is
happening in this case.

By the time the HMTL is compared with and without sanitisation in the
Validator, subtle whitespace changes have snuck which causes the `==`
operator to fail.

While we could definitely implement a bespoke fix for this issue, we've
decided to go with removing linebreaks from the html before comparing
it with and without sanitisation. It's largely just noise and could
definitely cause issues down the line again when further releases occur.