Disallow usage of style element and style attribute
This forbids a user from inputting a style element as part of their
govspeak markdown or using the style attribute in their HTML.
Thus it prevents publishers from customising the style of a GOV.UK page.

I am very confident we never intended to let users do this and consider
this a bug fix - luckily, the evidence of usage is minimal. This bug
provides publishers with an unprecedented ability to re-style a GOV.UK
page in a way far outside of the GOV.UK brand guidelines.

The reason I'm confident this was not intended is because we went to
special effort to only style on table cell elements 7 years ago [1]. It
appears that when we adopted Sanitize::Config::Relaxed in 2012 [2],
style was not allowed as it was later added to Sanitize::Config::Relaxed
in 2014 [3].

[1]: https://github.com/alphagov/govspeak/commit/b5b337e4424afb7e970f25e3c3843d35775db4e2
[2]: https://github.com/alphagov/govspeak/commit/c0fab20c3dc7cb38a6b768d9c2691c9653681a3d
[3]: https://github.com/rgrove/sanitize/commit/86d53f9adf65f5bbef99cbcd2322df478df9d59a