Add margin-bottom to embedded attachments
This change adds margin-bottom to embedded attachments. Without this,
attachments have zero margin-bottom, meaning they look visually
'squished up' when multiple attachments are embedded one after the other
in a document.

This styling makes it consistent with the 'Documents' section that
automatically gets added to some document types (for example,
Publications in Whitehall).