Add test for previously problematic scenario
In a previous Zendesk ticket
(https://govuk.zendesk.com/agent/tickets/4819383), a problem was
reported where abbreviations within HTML tag elements were being
replaced with abbreviation tags.

The fix at the time was to revert the code for adding abbreviations to
legislative lists and calls to action.

This adds a test for the previously problematic scenario, so we can
implement the abbreviations functionality again in a later commit.