Prevent acronyms inside acronyms
Currently, the `add_acronym_alt_text` runs through the page text for each
acronym. 

This means that if an acronym definition contains another acronym, the new
acronym will be nested inside the old, which is invalid HTML.

We have had several Zendesk tickets describing this problem.

This updates `add_acronym_alt_text` so that it only runs through the body once.