Parse call to action as part of the main conversion
Currently, we do a lot of work to replicate various functions of Kramdown
(acronyms and footnotes) in the call to action and legislative list
components.

The acronym function has been proved to contains bugs through several Zendesk
tickets.

1. Acronyms are inserted into the produced HTML, even if undesirable. For
example, a link `<email@example.com>` with the acronym `*[email]: Electronic
mail is a method of transmitting and receiving messages` would produce
`<p>href="mailto:<abbr title="Electronic mail is a method of transmitting and
receiving messages">email</abbr>@example.com"<abbr title="Electronic mail is a
method of transmitting and receiving messages">email</abbr>@example.com</p>`
rather than the expected link.

2. Because the `add_acronym_alt_text` method runs through the text for each
acronym, acronyms can be inserted into other acronyms creating invalid HTML.

Both of these issues can be prevented by parsing the call to action component
as part of the main conversion to HTML. Kramdown syntax is normally not
processed inside an HTML tag, but can be enabled with the
`parse_block_html` option.

There is a small regression in that the `$CTA` block must be proceeded by a
blank line (demonstrated in `test/govspeak_test.rb:428`), but this seems okay
as it's standard markdown.

Some of the tests can likely be removed as we are now testing the functionality
of the library, however, for now these represent that there are no regressions.